+++ app/e2e-tests/src/widgets/table.test.ts
@@ -4,27 +4,31 @@
 *--------------------------------------------------------------------------------------------*/
 import { Page } from "playwright";
 import { page } from "../setup";
-import { getEditor, getWidget, openTestIModel } from "../utils";
+import { getEditor, getStagePanelGrip, getWidget, openTestIModel } from "../utils";
 
-describe("properties widget #local", () => {
-  const contentUrlIdentifier = (url: URL) => {
-    return url.pathname.includes("getPagedContent");
+describe("table widget #local", () => {
+  const contentDescriptorUrlIdentifier = (url: URL) => {
+    return url.pathname.includes("getContentDescriptor");
   };
 
   beforeEach(async () => {
     await openTestIModel(page);
+
+    // expand the stage panel
+    const grip = getStagePanelGrip(page, "bottom");
+    await grip.dblclick();
   });
 
   afterEach(async () => {
-    await page.context().unroute(contentUrlIdentifier);
+    await page.context().unroute(contentDescriptorUrlIdentifier);
   });
 
   it("displays properties", async () => {
-    const propertiesWidget = getWidget(page, "Properties");
-    await propertiesWidget.locator("text=Select element(s) to view properties.").waitFor();
+    const tableWidget = getWidget(page, "Table");
+    await tableWidget.locator("text=Select element(s) to view properties.").waitFor();
 
     await selectAnyTreeNode(page);
-    await propertiesWidget.locator("text=Selected Item(s)").waitFor();
+    await tableWidget.locator("role=table").waitFor();
   });
 
   it("updates properties when ruleset changes", async () => {
@@ -34,27 +38,25 @@ describe("properties widget #local", () => {
     await page.click('text=""SelectedNodeInstances""');
     await editor.press("End");
     await editor.type(`,
-"propertyOverrides": [{ "name": "*", "categoryId": "custom" }],
-"propertyCategories": [{ "id": "custom", "label": "custom_category" }]`);
+"propertyOverrides": [{ "name": "Model", "labelOverride": "Custom Property Label" }]`);
     await editor.press("Alt+Enter");
 
-    const propertiesWidget = getWidget(page, "Properties");
-    await propertiesWidget.locator("text=custom_category").waitFor();
+    const tableWidget = getWidget(page, "Table");
+    await tableWidget.locator(`text="Custom Property Label"`).waitFor();
   });
 
-  // TODO: unskip after upgrading itwinjs-code dependencies to 4.0
-  it.skip("renders error status on error", async () => {
-    const propertiesWidget = getWidget(page, "Properties");
-    await propertiesWidget.locator("text=Select element(s) to view properties.").waitFor();
+  it("renders error status on error", async () => {
+    const tableWidget = getWidget(page, "Table");
+    await tableWidget.locator("text=Select element(s) to view properties.").waitFor();
 
     // simulate network error
     await page.context().route(
-      contentUrlIdentifier,
+      contentDescriptorUrlIdentifier,
       async (route) => route.abort(),
     );
 
     await selectAnyTreeNode(page);
-    await propertiesWidget.locator(`text="Error"`).waitFor();
+    await tableWidget.locator(`text="Error"`).waitFor();
   });
 
   // eslint-disable-next-line @typescript-eslint/no-shadow